tc: Remove unused error variants and rename `errors` to `diagnostics`
