+++ compiler/hash-typecheck/src/diagnostics/definitions.rs
@@ -1,6 +1,6 @@
+//! Definitions for errors and warnings that can occur during typechecking.
 use hash_ast::lit::LitParseError;
 use hash_reporting::diagnostic::IntoCompound;
-use hash_source::location::Span;
 use hash_tir::{
     tir::{
         fns::FnDefId, NodeOrigin, ParamError, ParamIndex, ParamsId, PatId, SomeParamsOrArgsId,
@@ -10,14 +10,14 @@ use hash_tir::{
 };
 use hash_utils::derive_more::From;
 
-/// A kind of wrong term usage.
+/// Some sort of wrong term usage.
 #[derive(Clone, Debug)]
 pub enum WrongTermKind {
     /// Cannot call a term because it is not a function.
     NotAFunction,
     /// Cannot access members because the term is not a record.
     ///
-    /// Records are tuples, single constructor data types.
+    /// Records are tuples and single constructor data types.
     NotARecord,
     /// Cannot index because the term is not an array.
     NotAnArray,
@@ -25,11 +25,16 @@ pub enum WrongTermKind {
     NotOfType { correct_ty: TyId },
 }
 
+/// The result type to be used for typechecking operations.
+pub type TcResult<T> = Result<T, TcError>;
+
 /// An error that occurs during typechecking.
 #[derive(Clone, Debug, From)]
-
 pub enum TcError {
     /// Blocked, cannot continue. This is used as a signal in the typechecker.
+    ///
+    /// @@Remove: this is to be removed once the typechecker can properly
+    /// handle unification blocking.
     Blocked(NodeOrigin),
 
     /// Signal to assert that there are other errors in the diagnostics store.
@@ -38,10 +43,10 @@ pub enum TcError {
     /// A series of errors.
     Compound { errors: Vec<TcError> },
 
-    /// More type annotations are needed to infer the type of the given term.
+    /// More type annotations are needed to infer the type of the given atom.
     NeedMoreTypeAnnotationsToInfer { atom: Atom },
 
-    /// More type annotations are needed to infer the type of the given term.
+    /// More type annotations are needed to unify the two atoms.
     NeedMoreTypeAnnotationsToUnify { src: Atom, target: Atom },
 
     /// The given arguments do not match the length of the target parameters.
@@ -50,45 +55,42 @@ pub enum TcError {
     /// The given parameters do not match the length of their annotations.
     WrongParamLength { given_params_id: ParamsId, annotation_params_id: ParamsId },
 
-    /// The two given argument lists cannot be unified due to mismatching
-    /// lengths.
+    /// The two given argument/parameter lists cannot be unified due to
+    /// mismatching lengths.
     DifferentParamOrArgLengths { a: SomeParamsOrArgsId, b: SomeParamsOrArgsId },
 
-    /// Cannot deref the subject.
+    /// Cannot dereference the subject as it is not a reference.
     CannotDeref { subject: TermId, actual_subject_ty: TyId },
 
-    /// Types don't match
+    /// Types don't match.
     MismatchingTypes { expected: TyId, actual: TyId },
 
-    /// Types don't match
+    /// Arrays have mismatching lengths.
     MismatchingArrayLengths { expected_len: TermId, got_len: TermId },
 
-    /// Wrong call kind
+    /// Wrong call kind, i.e. implicit/explicit.
     WrongCallKind { site: TermId, expected_implicit: bool, actual_implicit: bool },
 
-    /// Wrong type used somewhere
-    WrongTy { term: TermId, inferred_term_ty: TyId, kind: WrongTermKind },
+    /// Wrong term kind used somewhere, expected by a core language construct.
+    WrongTerm { term: TermId, inferred_term_ty: TyId, kind: WrongTermKind },
 
     /// The given property does not exist on the given term.
     PropertyNotFound { term: TermId, term_ty: TyId, property: ParamIndex },
 
-    /// Undecidable equality between terms
+    /// Undecidable equality between terms.
     UndecidableEquality { a: TermId, b: TermId },
 
-    /// Undecidable equality between terms
+    /// The two patterns do not match.
+    // @@Todo: remove this in favour of a single mismatch variant.
     MismatchingPats { a: PatId, b: PatId },
 
-    /// The given property does not exist on the given term.
+    /// The two functions do not match.
+    // @@Todo: remove this in favour of a single mismatch variant.
     MismatchingFns { a: FnDefId, b: FnDefId },
 
-    /// Invalid range pattern literal
-    InvalidRangePatternLiteral { location: Span },
-
-    /// Invalid range pattern literal
+    /// A type was returned from a block that references some non-substitutable
+    /// local variables in the block.
     TryingToReferenceLocalsInType { ty: TyId },
-
-    /// Cannot use the given term in a type position.
-    CannotUseInTyPos { location: NodeOrigin, inferred_ty: TyId },
 
     /// An error related to argument/parameter matching.
     #[from]
@@ -100,6 +102,7 @@ pub enum TcError {
     LitParseError(LitParseError),
 
     /// An error that occurred in an intrinsic.
+    /// This is a custom error message determined by the intrinsic.
     Intrinsic(String),
 }
 
@@ -108,5 +111,3 @@ impl IntoCompound for TcError {
         TcError::Compound { errors }
     }
 }
-
-pub type TcResult<T> = Result<T, TcError>;
