+++ compiler/hash-typecheck/src/diagnostics/mod.rs
@@ -1,0 +1,7 @@
+//! Diagnostic definitions and reporting, for errors that occur during
+//! typechecking.
+
+pub mod definitions;
+pub mod reporting;
+pub use definitions::*;
+pub use reporting::*;
