--- compiler/hash-typecheck/src/errors/mod.rs
@@ -1,5 +1,0 @@
-pub mod definitions;
-pub mod reporting;
-
-pub use definitions::*;
-pub use reporting::*;
