+++ compiler/hash-typecheck/src/operations/data.rs
@@ -12,8 +12,8 @@ use hash_tir::{
 };
 
 use crate::{
+    diagnostics::{TcError, TcResult},
     env::TcEnv,
-    errors::{TcError, TcResult},
     options::normalisation::{already_normalised, normalise_nested, NormaliseResult},
     tc::Tc,
     traits::{OperationsOn, OperationsOnNode, ScopedOperationsOnNode},
